requirements.txt
asgiref==3.7.2
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
Django==4.2.5
django-dotenv==1.4.2
dnspython==2.7.0
dominate==2.9.1
exceptiongroup==1.2.2
Flask==3.1.0
Flask-Bootstrap==3.3.7.1
gunicorn==21.2.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
outcome==1.3.0.post0
packaging==23.2
psycopg2-binary==2.9.9
pymongo==4.10.1
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.4.4
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
undetected-chromedriver==3.5.5
urllib3==2.3.0
visitor==0.1.3
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
wsproto==1.2.0
